+++ projects/packages/my-jetpack/_inc/admin.jsx
@@ -4,7 +4,7 @@
 import { ThemeProvider } from '@automattic/jetpack-components';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { createRoot } from '@wordpress/element';
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 import { HashRouter, Navigate, Routes, Route, useLocation } from 'react-router-dom';
 /**
  * Internal dependencies
@@ -17,7 +17,6 @@ import {
 	BackupInterstitial,
 	BoostInterstitial,
 	CRMInterstitial,
-	CreatorInterstitial,
 	ExtrasInterstitial,
 	JetpackAiInterstitial,
 	ProtectInterstitial,
@@ -70,7 +69,6 @@ const MyJetpack = () => {
 								<Route path={ MyJetpackRoutes.AddBackup } element={ <BackupInterstitial /> } />
 								<Route path={ MyJetpackRoutes.AddBoost } element={ <BoostInterstitial /> } />
 								<Route path={ MyJetpackRoutes.AddCRM } element={ <CRMInterstitial /> } />
-								<Route path={ MyJetpackRoutes.AddCreator } element={ <CreatorInterstitial /> } />
 								<Route
 									path={ MyJetpackRoutes.AddJetpackAI }
 									element={ <JetpackAiInterstitial /> }