+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/videopress-uploader/uploader-progress.js
@@ -209,6 +209,7 @@ const UploaderProgress = ( {
 				onChange={ newTitle => setAttributes( { title: newTitle } ) }
 				value={ title }
 				placeholder={ filename }
+				__nextHasNoMarginBottom={ true }
 			/>
 
 			<UploadingEditor