+++ projects/packages/yoast-promo/CHANGELOG.md
@@ -19,8 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Jetpack Autoloader package suggestion. [#29988]
 
 ### Changed
-- Updated package dependencies. [#29854]
-- Updated package dependencies. [#30019]
+- Updated package dependencies. [#29854] [#30019]
 
 ## 0.1.0 - 2023-04-03
 ### Added